+++ VKapp/VKapp/Models/FriendResult.swift
@@ -1,0 +1,9 @@
+// FriendResult.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Result of friend response
+struct FriendResult: Decodable {
+    let response: Friends
+}
