+++ VKapp/VKapp/Models/Friends.swift
@@ -1,0 +1,13 @@
+// Friends.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// List of friends
+struct Friends: Decodable {
+    let friends: [Friend]
+
+    enum CodingKeys: String, CodingKey {
+        case friends = "items"
+    }
+}
