+++ VKapp/VKapp/Models/Group.swift
@@ -3,20 +3,6 @@
 
 import Foundation
 
-/// Result of group response
-struct GroupResult: Decodable {
-    let response: Groups
-}
-
-/// List of groups
-struct Groups: Decodable {
-    let groups: [Group]
-
-    private enum CodingKeys: String, CodingKey {
-        case groups = "items"
-    }
-}
-
 /// Group
 final class Group: Decodable {
     @objc dynamic var nameGroup: String
