+++ VKapp/VKapp/Models/GroupResult.swift
@@ -1,0 +1,9 @@
+// GroupResult.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Result of group response
+struct GroupResult: Decodable {
+    let response: Groups
+}
