+++ VKapp/VKapp/Models/PhotoResponse.swift
@@ -1,0 +1,13 @@
+// PhotoResponse.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Result of photo response 2nd level
+struct PhotoResponse: Decodable {
+    let photoResponse: [Photos]
+
+    enum CodingKeys: String, CodingKey {
+        case photoResponse = "items"
+    }
+}
