+++ VKapp/VKapp/Models/PhotoResult.swift
@@ -1,0 +1,9 @@
+// PhotoResult.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Result of group response
+struct PhotoResult: Decodable {
+    let response: PhotoResponse
+}
